+++ src/analyze/+mag/@AutomatedAnalysis/AutomatedAnalysis.m
@@ -319,6 +319,38 @@ classdef (Sealed) AutomatedAnalysis < matlab.mixin.Copyable & mag.mixin.SetGet
                 findFinalNormalMode(this.Results.Secondary.Events, this.Results.Secondary.Time(end)));
         end
 
+        function periods = splitByTimeGap(this, gap)
+        % SPLITBYTIMEGAP Split data based on gap in the data of specified
+        % magnitude.
+
+            arguments (Input)
+                this
+                gap (1, 1) duration
+            end
+
+            arguments (Output)
+                periods (1, :) mag.Instrument
+            end
+
+            tPrimary = this.Results.Primary.Time;
+            dtPrimary = diff(tPrimary);
+            tSplitPrimary = [tPrimary(1); tPrimary(dtPrimary > gap)];
+
+            tSecondary = this.Results.Secondary.Time;
+            dtSecondary = diff(tSecondary);
+            tSplitSecondary = [tSecondary(1); tSecondary(dtSecondary > gap)];
+
+            if ~isequal(numel(tSplitPrimary), numel(tSplitSecondary))
+                error("Unequal time splits in primary (%d) and secondary(%d) data. Try a different time gap.", numel(tSplitPrimary), numel(tSplitSecondary));
+            end
+
+            for i = 1:(numel(tSplitPrimary) - 1)
+
+                periods(i) = this.applyTimeRangeToInstrument(timerange(tSplitPrimary(i), tSplitPrimary(i + 1), "open"), ...
+                    timerange(tSplitSecondary(i), tSplitSecondary(i + 1), "open")); %#ok<AGROW>
+            end
+        end
+
         % EXPORT Export data to specified format.
         export(this, exportStrategy, options)
     end